error after adding four fields

+++ lib/view/notes/create_update_note_view.dart
@@ -17,12 +17,18 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   CloudNote? _note;
   late final FirebaseCloudStorage _notesService;
   late final TextEditingController _textController;
+  late final TextEditingController _textController1;
+  late final TextEditingController _textController2;
+  late final TextEditingController _textController3;
 
   Future<CloudNote> createOrGetExistingNote(BuildContext context) async {
     final widgetNote = context.getArgument<CloudNote>();
     if (widgetNote != null) {
       _note = widgetNote;
       _textController.text = widgetNote.text;
+      _textController1.text = widgetNote.text;
+      _textController2.text = widgetNote.text;
+      _textController3.text = widgetNote.text;
       return widgetNote;
     }
 
@@ -39,7 +45,11 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
 
   void _deleteNoteIfTextIsEmpty() {
     final note = _note;
-    if (_textController.text.isEmpty && note != null) {
+    if (_textController.text.isEmpty &&
+        _textController1.text.isEmpty &&
+        _textController2.text.isEmpty &&
+        _textController3.text.isEmpty &&
+        note != null) {
       _notesService.deleteNotes(documentId: note.documentId);
     }
   }
@@ -47,10 +57,20 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   void _saveNoteIfTextNotEmpty() async {
     final note = _note;
     final text = _textController.text;
-    if (note != null && text.isNotEmpty) {
+    final text1 = _textController1.text;
+    final text2 = _textController2.text;
+    final text3 = _textController3.text;
+    if (note != null &&
+        (text.isNotEmpty ||
+            text1.isNotEmpty ||
+            text2.isNotEmpty ||
+            text3.isNotEmpty)) {
       await _notesService.updateNote(
         documentId: note.documentId,
-        text: text,
+          text: text,
+          text1: text1,
+          text2: text2,
+          text3: text3,
       );
     }
   }
@@ -60,6 +80,9 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
     _deleteNoteIfTextIsEmpty();
     _saveNoteIfTextNotEmpty();
     _textController.dispose();
+    _textController1.dispose();
+    _textController2.dispose();
+    _textController3.dispose();
     super.dispose();
   }
 
@@ -67,6 +90,9 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   void initState() {
     _notesService = FirebaseCloudStorage();
     _textController = TextEditingController();
+    _textController1 = TextEditingController();
+    _textController2 = TextEditingController();
+    _textController3 = TextEditingController();
     super.initState();
   }
 
@@ -76,9 +102,15 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
       return;
     }
     final text = _textController.text;
-    await _notesService.updateNote(
-      documentId: note.documentId,
+    final text1 = _textController1.text;
+    final text2 = _textController2.text;
+    final text3 = _textController3.text;
+    await _notesService
+        .updateNote(documentId: note.documentId, 
       text: text,
+      text1: text1,
+      text2: text2,
+      text3: text3,
     );
   }
 
@@ -96,10 +128,20 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
           IconButton(
             onPressed: () async {
               final text = _textController.text;
-              if (_note == null || text.isEmpty) {
+              final text1 = _textController1.text;
+              final text2 = _textController2.text;
+              final text3 = _textController3.text;
+              if (_note == null ||
+                  (text.isEmpty &&
+                      text1.isEmpty &&
+                      text2.isEmpty &&
+                      text3.isEmpty)) {
                 await showCannotShareEmptyNoteDialog(context);
               } else {
                 Share.share(text);
+                Share.share(text1);
+                Share.share(text2);
+                Share.share(text3);
               }
             },
             icon: const Icon(Icons.share),
@@ -121,25 +163,25 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                     decoration: const InputDecoration(
                       hintText: "Enter diagnosis",
                     ),
-                  ),  
+                  ),
                   TextField(
-                    controller: _textController,
+                    controller: _textController1,
                     keyboardType: TextInputType.multiline,
                     maxLines: null,
                     decoration: const InputDecoration(
                       hintText: "Enter date of diagnosis",
                     ),
-                  ),  
+                  ),
                   TextField(
-                    controller: _textController,
+                    controller: _textController2,
                     keyboardType: TextInputType.multiline,
                     maxLines: null,
                     decoration: const InputDecoration(
                       hintText: "Enter prescriptions",
                     ),
-                  ),  
+                  ),
                   TextField(
-                    controller: _textController,
+                    controller: _textController3,
                     keyboardType: TextInputType.multiline,
                     maxLines: null,
                     decoration: const InputDecoration(
